+++ app/src/main/java/ru/practicum/android/diploma/di/InteractorModule.kt
@@ -26,6 +26,7 @@ val interactorModule = module {
     factory<FilterInteractor> {
         FilterInteractorImpl(
             repository = get(),
+            repositorySharePrefs = get()
         )
     }
 }
